+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.2.3
 ###### January 16, 2024
+- Notified of a new error log if a tileset texture's texture is not a TextureType of Sprite.
 - Fixed a problem where the export app would not work on Macs 
   - This issues is not retroactively fixed, so for any Mac users currently facing this issue, add a "$1" to the end of the file's contents in `Library/LDtkTilesetExporter/ExportTilesetDefinitionMax.sh`)